Bug fix in couch_work_queue: state's size not decremented
When dequeing some items from the queue (that is, not taking all the queued items)
the size field of the gen_server's state was not being decremented as it should.
However when all the queue items are delivered to a consumer, the size is set to 0.

This particular fix has a good impact in the new replicator because it uses work queues
and the consumers only dequeue 1 item at a time - producers will no longer be blocked
until the queue gets empty.
For a push replication that used to take about 15 minutes, it now takes about 13 minutes.



git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1103266 13f79535-47bb-0310-9956-ffa450edef68