Simpler and safer db open/closing in view group servers
This makes the opening and closing of databases in the view
group server to be more friendly with the db reference counting
system, avoiding more potential db file leaking after compaction,
as we currently open a database in one process and use it on
another process (view compactor, view updater).

This significantly reduces the chances of failure when compacting
very large views as discussed in COUCHDB-994.

This relates to COUCHDB-926 and COUCHDB-994.

git-svn-id: https://svn.apache.org/repos/asf/couchdb/trunk@1138796 13f79535-47bb-0310-9956-ffa450edef68

Conflicts:

	src/couchdb/couch_view_compactor.erl
	src/couchdb/couch_view_group.erl
	src/couchdb/couch_view_updater.erl

Change-Id: I99a029aa29449c17509799f41bec9bea5c07b937
Reviewed-on: http://review.couchbase.org/7146
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>