Add `since_seq` parameter to replication objects
This optional parameter allows a user to start a replication
from a specified source database sequence number (similar to
the ?since=SeqNum parameter for /db/_changes/). The advantage
of this is that to replicate very large databases, the user
can copy the database file into another server (via rsync for
example) and then start a continuous replication with the
"since_seq" parameter set to the value of the copy's update
sequence number, speeding up the replication process.

Change-Id: Ia58c4e032d99aee161caf17daad46f345c99bec0
Reviewed-on: http://review.couchbase.org/7150
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Dustin Sallings <dustin@spy.net>
Tested-by: Dustin Sallings <dustin@spy.net>