Reduce the possibility of checkpoint conflicts
If a replication checkpoint update succeeds but we receive
a timeout before receiving the success response, the update
is retried and it will eventually fail with a conflict error.
When the conflict error happens, just grab the current
checkpoint revision from the endpoint and use it for the
next checkpoint update.

Change-Id: Ibdd71d2acebc692e2957ecc0686e1ba4452f9d7a
Reviewed-on: http://review.couchbase.org/7599
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Dustin Sallings <dustin@spy.net>